Test.Generators: module ByteString generators (and its tests)
This prepares for Test.Async.Socket where these generators will also be
useful.
